Enable simultaneous cpp and C# project compilation
The original compilation guide of the Antenna software is defined by the
sln and csproj files. I cannot find any method to covert them into a
CMake project so that I can use CMakeLists to instruct the compilation
of it. Hence, I added some custom compilation command in the top level
CMake file so that it will automatically call the msbuild to compile
them according to its origial csproj file.